third_party/blink/web_tests/virtual/font-variant-emoji/external/wpt/css/css-fonts/font-shorthand-subproperties-reset-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-variant-emoji should be reset to its initial value.
  assert_equals: expected "normal" but got "text"
Harness: the test ran to completion.
